Refactor: Relocate `BankAppBaseException` to shared.exception package
