+++ src/main/java/info/mackiewicz/bankapp/shared/exception/IbanAnalysisException.java
@@ -1,6 +1,5 @@
 package info.mackiewicz.bankapp.shared.exception;
 
-import info.mackiewicz.bankapp.system.error.handling.core.BankAppBaseException;
 import info.mackiewicz.bankapp.system.error.handling.core.error.ErrorCode;
 
 public class IbanAnalysisException extends BankAppBaseException {
