+++ src/main/java/info/mackiewicz/bankapp/system/error/handling/mapping/ApiExceptionToErrorMapper.java
@@ -2,7 +2,7 @@ package info.mackiewicz.bankapp.system.error.handling.mapping;
 
 import org.springframework.stereotype.Component;
 
-import info.mackiewicz.bankapp.system.error.handling.core.BankAppBaseException;
+import info.mackiewicz.bankapp.shared.exception.BankAppBaseException;
 import info.mackiewicz.bankapp.system.error.handling.core.error.ErrorCode;
 
 /**
