+++ src/main/java/info/mackiewicz/bankapp/system/registration/exception/DemoRegistrationException.java
@@ -1,6 +1,6 @@
 package info.mackiewicz.bankapp.system.registration.exception;
 
-import info.mackiewicz.bankapp.system.error.handling.core.BankAppBaseException;
+import info.mackiewicz.bankapp.shared.exception.BankAppBaseException;
 import info.mackiewicz.bankapp.system.error.handling.core.error.ErrorCode;
 
 public class DemoRegistrationException extends BankAppBaseException {
